# proyecto-dws-grupo2/api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: TitanGym API
  description: Documentación oficial de la API REST del gimnasio TitanGym
  version: "1.0"
servers:
- url: https://localhost:8443
  description: Generated server url
security:
- bearerAuth: []
...
paths:
  /api/v1/users/{id}:
    get:
      tags:
      - Usuarios
      summary: Obtener usuario por ID
      description: Busca la información de un miembro específico mediante su identificador.
      operationId: getUserById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: No existe ningún usuario con ese ID
          content: {}
        "200":
          description: Usuario encontrado
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
    put:
      tags:
      - Usuarios
      summary: Actualizar un usuario
      description: "Modifica los datos de un miembro existente. Si se envía una contraseñ\
        a, será encriptada."
      operationId: updateUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRequestDTO"
        required: true
      responses:
        "404":
          description: Usuario no encontrado
          content: {}
        "200":
          description: Usuario actualizado correctamente
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
        "400":
          description: Datos de actualización inválidos
    delete:
      tags:
      - Usuarios
      summary: Eliminar un usuario
      description: Borra permanentemente la cuenta de un miembro.
      operationId: deleteUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Usuario no encontrado
          content: {}
        "200":
          description: Usuario eliminado con éxito
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
  /api/v1/services/{id}:
    get:
      tags:
      - Servicios
      summary: Obtener un servicio por su ID
      description: Busca la información detallada de un servicio específico.
      operationId: getService
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: No existe ningún servicio con ese ID
          content: {}
        "200":
          description: Servicio encontrado
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ServiceResponseDTO"
    put:
      tags:
      - Servicios
      summary: Actualizar un servicio
      description: Modifica los datos de un servicio ya existente en el sistema.
      operationId: updateService
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceRequestDTO"
        required: true
      responses:
        "404":
          description: Servicio no encontrado para actualizar
          content: {}
        "200":
          description: Servicio actualizado correctamente
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ServiceResponseDTO"
        "400":
          description: Datos de actualización inválidos
    delete:
      tags:
      - Servicios
      summary: Eliminar un servicio
      description: Borra permanentemente un servicio del catálogo.
      operationId: deleteService
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Servicio no encontrado para eliminar
          content: {}
        "200":
          description: Servicio eliminado con éxito
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ServiceResponseDTO"
...
  /api/v1/users/:
    post:
      tags:
      - Usuarios
      summary: Registrar un nuevo usuario
      description: Crea una nueva cuenta de usuario en el sistema.
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRequestDTO"
        required: true
      responses:
        "404":
          description: Not Found
        "201":
          description: Usuario creado con éxito
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
        "400":
          description: Datos de registro inválidos
...
  /api/v1/services/:
    get:
      tags:
      - Servicios
      summary: Obtener todos los servicios
      description: "Devuelve una página con los servicios disponibles (Fisioterapia,\
        \ Nutrición, Bebidas, etc.)."
      operationId: getServices
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "404":
          description: Not Found
        "200":
          description: Catálogo de servicios recuperado con éxito
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedModelServiceResponseDTO"
    post:
      tags:
      - Servicios
      summary: Crear un nuevo servicio
      description: Añade un nuevo servicio al catálogo general.
      operationId: createService
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceRequestDTO"
        required: true
      responses:
        "404":
          description: Not Found
        "201":
          description: Servicio creado con éxito
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ServiceResponseDTO"
        "400":
          description: Los datos del servicio son incorrectos
...
  /api/v1/users:
    get:
      tags:
      - Usuarios
      summary: Obtener todos los usuarios
      description: Devuelve una página con la lista de todos los miembros registrados.
        Requiere permisos de administrador.
      operationId: getAllUsers
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "404":
          description: Not Found
        "200":
          description: Lista de usuarios recuperada con éxito
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedModelUserResponseDTO"
  /api/v1/users/me:
    get:
      tags:
      - Usuarios
      summary: Obtener mi perfil
      description: Devuelve los datos del usuario que ha iniciado sesión actualmente.
      operationId: getMe
      responses:
        "404":
          description: Not Found
        "200":
          description: Perfil recuperado con éxito
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
        "401":
          description: No autorizado / No hay sesión activa
...
components:
  schemas:
    UserRequestDTO:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        password:
          type: string
    ImageResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        url:
          type: string
    UserResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        roles:
          type: array
          items:
            type: string
        profileImage:
          $ref: "#/components/schemas/ImageResponseDTO"
    ServiceRequestDTO:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        serviceKey:
          type: string
    ServiceResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        serviceKey:
          type: string
        enrolled:
          type: boolean
        image:
          $ref: "#/components/schemas/ImageResponseDTO"
...
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
    PageMetadata:
      type: object
      properties:
        size:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
    PagedModelUserResponseDTO:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/UserResponseDTO"
        page:
          $ref: "#/components/schemas/PageMetadata"
    PagedModelServiceResponseDTO:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/ServiceResponseDTO"
        page:
          $ref: "#/components/schemas/PageMetadata"
...
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT